refactor: split MCP server into focused modules
