+++ pyproject.toml
@@ -12,6 +12,7 @@ dependencies = [
     "mcp[cli]==1.25.0",
     "orjson>=3.11.1",
     "certifi>=2024.0.0",
+    "httpx>=0.27.0",
     "PyJWT[crypto]>=2.9.0",
 ]
